Adjust comment endpoints

+++ Backend/MaternityCare_Backend/MaternityCare_Backend.Service/CommentServices/CommentService.cs
@@ -18,13 +18,11 @@ namespace MaternityCare_Backend.Service.CommentServices
 			this.mapper = mapper;
 		}
 
-		private async Task<Comment?> CheckCommentExist(Guid blogId, Guid userId, Guid commentId, bool trackChange, CancellationToken ct = default)
+		private async Task<Comment?> CheckCommentExist(Guid userId, Guid commentId, bool trackChange, CancellationToken ct = default)
 		{
-			var blogEntity = await repositoryManager.BlogRepository.GetBlog(blogId, false, ct);
-			if (blogEntity == null) throw new BlogNotFoundException();
 			var userEntity = await repositoryManager.UserRepository.GetUserById(userId, false, ct);
 			if (userEntity == null) throw new UserNotFoundException();
-			var commentEntity = await repositoryManager.CommentRepository.GetComment(blogId, userId, commentId, trackChange, ct);
+			var commentEntity = await repositoryManager.CommentRepository.GetComment(commentId, trackChange, ct);
 			if (commentEntity == null) throw new CommentNotFoundException();
 			return commentEntity;
 		}
@@ -39,9 +37,10 @@ namespace MaternityCare_Backend.Service.CommentServices
 			await repositoryManager.SaveAsync(ct);
 		}
 
-		public async Task DeleteComment(Guid blogId, Guid userId, Guid commentId, CancellationToken ct = default)
+		public async Task DeleteComment(Guid userId, Guid commentId, CancellationToken ct = default)
 		{
-			var commentEntity = await CheckCommentExist(blogId, userId, commentId, false, ct);
+			var commentEntity = await CheckCommentExist(userId, commentId, false, ct);
+			if (commentEntity.UserId != userId) throw new CommentConflictException("You can not delete other people comments");
 			repositoryManager.CommentRepository.DeleteComment(commentEntity);
 			await repositoryManager.SaveAsync(ct);
 		}
@@ -53,9 +52,10 @@ namespace MaternityCare_Backend.Service.CommentServices
 			return (comments, commentsWithMetaData.MetaData);
 		}
 
-		public async Task UpdateComment(Guid blogId, Guid userId, Guid commentId, CommentForUpdateDto commentForUpdateDto, CancellationToken ct = default)
+		public async Task UpdateComment(Guid userId, Guid commentId, CommentForUpdateDto commentForUpdateDto, CancellationToken ct = default)
 		{
-			var commentEntity = await CheckCommentExist(blogId, userId, commentId, true, ct);
+			var commentEntity = await CheckCommentExist(userId, commentId, true, ct);
+			if (commentEntity.UserId != userId) throw new CommentConflictException("You can not modify other people comments");
 			mapper.Map(commentForUpdateDto, commentEntity);
 			commentEntity.UpdatedAt = DateTime.Now;
 			await repositoryManager.SaveAsync(ct);
